requirements.txt
playwright==1.46.0
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
psycopg2-binary==2.9.10
fastapi==0.115.4
uvicorn==0.32.0
requests==2.32.3
SQLAlchemy==2.0.32
PyJWT==2.9.0
python-multipart==0.0.20
